Routing and Navigation
To route or nest. When routing, selector is not required.
Routing requires RouterModule. Configure the routes, base element in html file, add RouterModule.forRoot.
Each route requires path. empty quotes for default route. Wildcard route provided by '**'.
Tie routes to actions - routerLink attribute on hyperlink elements.
When a routerLink is chosen, it has to be rendered in the routerOutlet.